Extract shared alignment core from the interactive service backend
Lift the matching and fitting logic out of interactive_alignment.py into
alignment_core.py so the upcoming align_cameras pipeline process and the
interactive service share one implementation. The stdio protocol and its
behavior are unchanged; InteractiveAlignmentService now delegates to a
LoftrMatcher plus register_image_pair.

The 16-bit percentile stretch window is promoted from a hardcoded module
constant to lower_percentile/upper_percentile options (with per-camera
cam2_* overrides), preserving the invariant that 8-bit imagery is never
stretched: bit depth decides, not configuration.
