+++ plugins/core/CMakeLists.txt
@@ -487,6 +487,11 @@ if( VIAME_ENABLE_PYTHON )
     interactive_stereo )
 
   kwiver_add_python_module(
+    ${CMAKE_CURRENT_SOURCE_DIR}/alignment_core.py
+    core
+    alignment_core )
+
+  kwiver_add_python_module(
     ${CMAKE_CURRENT_SOURCE_DIR}/interactive_alignment.py
     core
     interactive_alignment )
